# PERF/CLN: Use Numpy C Iterator API to unify isnaobj 1D/2D