@mroeschke 2023-01-13
```
mypy (stubtest)..........................................................Failed
- hook id: stubtest
- duration: 94.59s
- exit code: 1

pandas._libs.missing.isnaobj2d is not present at runtime
```